+++ Assets/DeltaDNA/DDNA.cs
@@ -493,10 +493,17 @@ namespace DeltaDNA
         public AndroidNotifications AndroidNotifications { get; private set; }
 
         /// <summary>
-        /// Clears the persistent data such as user id.  Useful for testing purposes.
+        /// Clears the persistent data, such as user id. The SDK should be stopped
+        /// before this method is called.
+        /// 
+        /// Useful for testing purposes.
         /// </summary>
         public void ClearPersistentData()
         {
+            if (HasStarted) {
+                Logger.LogWarning("SDK has not been stopped before clearing persistent data");
+            }
+
             PlayerPrefs.DeleteKey(PF_KEY_USER_ID);
 
             if (this.eventStore != null) {